+++ changelog.md
@@ -2,6 +2,7 @@
 - Refactored to use `Invoke-OpenAIAPI` function. This function is used by all the other functions in the module. This allows for a single place to update the API URL and the API Key. 
 - Add `Get-*` functions for OpenAI URIs
 - Took the function suggestions from [Skatterbrainz](https://github.com/Skatterbrainz) and updated with `Invoke-OpenAIAPI`  the refactor: https://github.com/dfinke/PowerShellAI/pull/30
+- Refactored `Get-DalleImage` to use `Invoke-OpenAIAPI`
 
 # v0.3.3
 - Check if `$result.choices` is not null before trying to access it. Thank you [StartAutomating](https://github.com/StartAutomating)